pyproject.toml
[tool.poetry]
name = "assignment1"
version = "0.1.0"
description = ""
authors = ["Alexander Talbott <atalbott13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
CensusData = "^1.15"
pandas = "^1.3.2"
PyYAML = "^5.4.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
